refactoring components + remove console.logs
