+++ src/App.js
@@ -18,6 +18,10 @@ import {
 import ProductItem from "./components/ProductItem";
 import ProductPage from "./components/ProductPage";
 import ContactPage from "./components/ContactPage";
+import { CheckOutPage } from "./components/CheckoutPage";
+import { CheckOutSuccessPage } from "./components/CheckoutSuccessPage";
+import { Cart } from "./components/Cart";
+import { Search } from "./components/Search";
 import * as Styled from "./App.styles";
 import CartIcon from "./images/cart.svg";
 
@@ -56,53 +60,6 @@ function Home() {
         <ProductItem key={item.id} data={item} />
       ))}
     </Styled.ProductListContainer>
-  );
-}
-
-function CheckOutPage() {
-  const { state } = useContext(CartContext);
-  const { cart, total, quantity } = state;
-
-  console.log(state);
-
-  console.log(cart);
-  return (
-    <Styled.CheckoutContainer>
-      {cart.length > 0 ? (
-        <>
-          <ul>
-            {cart.map((item) => (
-              <li key={item.id}>
-                <img src={item.imageUrl} />
-                <p>{item.title} - </p>
-                <p>Qty: {item.quantity}</p>
-              </li>
-            ))}
-            <h2>Total: {total.toFixed(2)}</h2>
-          </ul>
-          <ClearCart />
-        </>
-      ) : (
-        <>
-          <h1>Cart is empty</h1>
-          <Styled.Button>
-            <Link to={`/`}>Return to home</Link>
-          </Styled.Button>
-        </>
-      )}
-    </Styled.CheckoutContainer>
-  );
-}
-
-function CheckOutSuccessPage() {
-  const [state, dispatch] = useReducer(reducer, initialState);
-  return (
-    <Styled.CheckoutSuccess>
-      <h1>Checkout was successful</h1>
-      <Styled.Button>
-        <Link to={`/`}>Return to home</Link>
-      </Styled.Button>
-    </Styled.CheckoutSuccess>
   );
 }
 
@@ -184,9 +141,6 @@ export function reducer(state, action) {
         currentTotal += product.discountedPrice * product.quantity;
         return currentTotal;
       }, 0);
-      console.log("New Total:", newTotal);
-      console.log("New Qty:", newQuantity);
-      console.log(state);
       return { ...state, cart: cart, total: newTotal, quantity: newQuantity };
 
     case "clearCart":
@@ -234,64 +188,6 @@ export function ClearCart() {
     navigate("/checkoutsuccess");
   };
   return <Styled.Button onClick={handleClear}>Checkout</Styled.Button>;
-}
-
-function Cart() {
-  const { state } = useContext(CartContext);
-  const { cart, total, quantity } = state;
-
-  return (
-    <div>
-      <Link to="/checkout">
-        <img className="cart-icon" src={CartIcon} />
-      </Link>
-      <p className="counter"> {quantity} </p>
-    </div>
-  );
-}
-
-function Search(props) {
-  const [searchQuery, setSearchQuery] = useState("");
-  const [searchResults, setSearchResults] = useState([]);
-
-  const handleSearch = (event) => {
-    event.preventDefault();
-    const results = props.data.filter((item) => {
-      return item.title.toLowerCase().includes(searchQuery.toLowerCase());
-    });
-    setSearchResults(results);
-  };
-
-  return (
-    <div>
-      <Styled.SearchFormContainer>
-        <form id="search-form" onKeyUp={handleSearch}>
-          <label htmlFor="search"></label>
-          <input
-            name="search"
-            id="search"
-            placeholder="Search..."
-            value={searchQuery}
-            onChange={(event) => setSearchQuery(event.target.value)}
-          />
-        </form>
-      </Styled.SearchFormContainer>
-      <div>
-        {searchResults.map((item) => (
-          <Styled.ProductContainer key={item.id}>
-            <li>
-              <h2>{item.title}</h2>
-              <img src={item.imageUrl} alt={item.title} />
-              <p>{item.description}</p>
-              <Styled.Button>
-                <Link to={`/product/${item.id}`}>View product</Link>
-              </Styled.Button>
-            </li>
-          </Styled.ProductContainer>
-        ))}
-      </div>
-    </div>
-  );
 }
 
 function App() {
