+++ src/components/CheckoutSuccessPage/index.jsx
@@ -1,0 +1,16 @@
+import { useReducer } from "react";
+import { reducer, initialState } from "../../App";
+import { Link } from "react-router-dom";
+import * as Styled from "../../App.styles";
+
+export function CheckOutSuccessPage() {
+  const [state, dispatch] = useReducer(reducer, initialState);
+  return (
+    <Styled.CheckoutSuccess>
+      <h1>Checkout was successful</h1>
+      <Styled.Button>
+        <Link to={`/`}>Return to home</Link>
+      </Styled.Button>
+    </Styled.CheckoutSuccess>
+  );
+}
